code/patrick/react/capstone/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1
django-crispy-forms==1.7.2
idna==2.7
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
asgiref==3.4.1
autopep8==1.5.7
certifi==2018.10.15
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
crispy-bootstrap5==0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-feedparser==0.2.1
djangorestframework==3.12.4
feedparser==6.0.8
idna==2.7
oauthlib==3.1.1
Pillow==8.3.2
prettytable==2.2.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.23
wcwidth==0.2.5
xmltodict==0.12.0